Special case Throne Room logic - Feast doesn't get trashed twice.  You still gain 2 cards, though, so it's not altogether different (old code was duping Feast cards in the Trash).
Fixed a bug where trashing could trash cards in play if they weren't in hand, which likely never happened anyway.

Fixed a bug where AI would try to gain cards that weren't left in the supply.  This makes the Random players not try to gain Duchies and fail all the time.